server/ResearchAgent/requirements.txt
requests
beautifulsoup4
feedparser
scheduler
sentence-transformers
chromadb
groq
python-dotenv
yfinance
redis
aiohttp
pyrate-limiter
asyncio
ntscraper
numpy
transformers
pinecone
hashlib
